Rename provider folder
