+++ packages/logging/src/otelProvider/internal-types.ts
